Show OpenSource section only in one place
- fix localization
