+++ Packages/BTLBSettings/Sources/BTLBSettings/Resources/Localizable.xcstrings
@@ -125,6 +125,16 @@
         }
       }
     },
+    "GitHub Repository" : {
+      "localizations" : {
+        "de" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "GitHub Repository"
+          }
+        }
+      }
+    },
     "INFO_VIEW_LIBRARIES_TEXT" : {
       "localizations" : {
         "de" : {
@@ -767,6 +777,26 @@
         }
       }
     },
+    "License" : {
+      "localizations" : {
+        "de" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Lizenz"
+          }
+        }
+      }
+    },
+    "Links" : {
+      "localizations" : {
+        "de" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Links"
+          }
+        }
+      }
+    },
     "NOTIFICATION_ACCOUNT_LAST_AUTOMATIC_UPDATE" : {
       "extractionState" : "manual",
       "localizations" : {
@@ -909,6 +939,32 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Notifications"
+          }
+        }
+      }
+    },
+    "Open Source" : {
+      "localizations" : {
+        "de" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Open Source"
+          }
+        }
+      }
+    },
+    "OPEN_SOURCE_DESCRIPTION" : {
+      "localizations" : {
+        "de" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Im Jahr 2025, 15 Jahre nach seiner ersten Veröffentlichung, wurde \nBTLB | bɪtlɪb | endlich Open Source 🦄.\n\nDieser Abschnitt ist den Stunden, der Leidenschaft und der Kreativität gewidmet, die jeder in diese Anwendung investiert hat. Danke, dass ihr BTLB besser gemacht habt, als es ohne euch jemals hätte sein können 💖"
+          }
+        },
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "In 2025, 15 years after its first release, \nBTLB | bɪtlɪb | finally became Open Source 🦄\n\nThis section is dedicated to everyone’s hours, passion and creativity spent on this application. Thank you for making BTLB better than it could have ever been without you 💖"
           }
         }
       }
