--- Packages/BTLBSettings/Sources/BTLBSettings/Sections/About Section/OpenSourceSectionCoordinator.swift
@@ -1,15 +1,0 @@
-import SwiftUI
-
-import ArchitectureX
-
-public class OpenSourceSectionCoordinator: Coordinator {
-    public var router: Router?
-    
-    public init(router: Router? = nil) {
-        self.router = router
-    }
-    
-    public var contentView: some View {
-        OpenSourceSectionView()
-    }
-}
